my version of ticketviewer for now
well i couldnt get the addArray at the end of statement ot work it just
greys out but the sarray is set up in create ticket and the listview is
set up in the same window as createticket and the view ticket is now
going to the createticket via changing the buttons in
appcompatpreferenceeactivity.   BUT the FAQ is fully functional too bad
lost a lot of time on errors.